chore: separate artwork card meta
